reverted back

+++ src/main/java/com/hbng/miniapp/charger/model/Customer.java
@@ -6,7 +6,6 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import java.math.BigDecimal;
 
 @Entity
 @Getter @Setter @NoArgsConstructor @AllArgsConstructor @ToString
@@ -15,9 +14,9 @@ public class Customer {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String accountNumber;
-    private BigDecimal transactedAmount;
+    private double transactedAmount;
 
-    public Customer(String accountNumber, BigDecimal transactedAmount) {
+    public Customer(String accountNumber, double transactedAmount) {
         this.accountNumber = accountNumber;
         this.transactedAmount = transactedAmount;
     }
