+++ src/main/java/com/hbng/miniapp/charger/services/CustomerService.java
@@ -2,7 +2,6 @@ package com.hbng.miniapp.charger.services;
 
 import com.hbng.miniapp.charger.model.Customer;
 
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
@@ -10,12 +9,12 @@ public interface CustomerService {
 
      List<Customer> findAll();
 
-     List<Customer> findByAmount(BigDecimal amount);
+     List<Customer> findByAmount(double amount);
      Customer findById(Long theId);
 
      void save(Customer customer);
 
      void deleteById(Long theId);
 
-     Map<String, Integer> myMap(BigDecimal amount);
+     Map<String, Integer> myMap(double amount);
 }
